- more fixup for math functions - more performance optimizations - remove gcem due to inconsistent result with the cmath funcs - fixup more ares stuffs - fix bug where team just produced from warfactory will go thru edge of the map - fix bug where wrong function was used on TemporalClass::_AI backport - fix IonCannon AI targeting was not listed to default targeting configurations - fix bug were some TEvent selecting wrong house - fix the FindAndTakeVehicle algorithm to match ares - fix CustomFoundation parsing style to match original ares - fix BombClass frame selecting to match original ares - fix ftol replacement was not suitable for certain value causing bug - fixup the cell_Distance_Squared to match original ares - fixup some more enum class to magic_enum for easy use - improve sorting algorithm with pdqshort - sync to some to develop - add /Zc:__cplusplus for enabling appropriate features that using the macro

+++ src/Misc/Ares/Hooks/Hooks.Strips.cpp
@@ -830,8 +830,7 @@ ASMJIT_PATCH(0x6AAD2F, SelectClass_ProcessInput_LoadCameo1, 7)
 
 	auto& cameos = MouseClassExt::TabCameos[MouseClass::Instance->ActiveTabIndex];
 
-	if (CameoIndex >= cameos.size())
-	{
+	if ((size_t)CameoIndex >= cameos.size()) {
 		return 0x6AB94F;
 	}
 
@@ -1092,14 +1091,14 @@ ASMJIT_PATCH(0x6aa600, StripClass_RecheckCameos, 5)
 
 	tabs.erase(iter, tabs.end());
 
-	if (tabs.size() >= pThis->BuildableCount)
+	if ((int)tabs.size() >= pThis->BuildableCount)
 	{
 		R->EAX(0);
 		return 0x6AACAE;
 	}
 
 	pThis->BuildableCount = tabs.size();
-	if (tabs.size() <= 0)
+	if ((int)tabs.size() <= 0)
 	{
 		SidebarClass::ShapeButtons[pThis->TabIndex].Disable();
